requirements.txt
fastapi
uvicorn
httpx
python-dotenv
pydantic
langchain
sqlalchemy
asyncpg
openai